Improve node APIs
This patch is based on r3471 with the following changes: 
+ Has the build-break fixes from r3472 
+ Fixes the _Native* class name issue that caused the new frog compiler 
to fail. (Note, the new frog compiler hasn't been checked in yet, so I'm 
not sure all the errors are fixed.) 
+ Incorporates all review feedback for r3471 

Put each node module in its own dart library. 

Add support for several node modules. 

+ os 
+ net 
+ dns 
+ http 
+ url 
+ querystring 
+ cluster 
+ zlib 
+ Buffer 
+ typed arrays 
+ crypto 

Create interfaces for the different types of streams, 
so that readable, writable, and readWrite streams 
all share protocol. 

Add simple echoServer and httpServer samples. 

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@3519 260f80e4-7a28-3924-810f-c04153c831b5