Visual emphasis for resources defining libraries in the Files View.
As per our conversation at the summit, this emphasizes libraries[1] in the Files View by making them bold. 


[1] Actually, to be more precise, this calls out "analysis roots", a more loose definition of library.  More context from Brian on that:

"We (no I don't remember who all was involved in the decision) decided awhile ago to define a library to be any file that is not known to be #source'd into another library. There were two primary rationales that I recall. First, while a library is required to contain a #library directive, a script is not, and we probably want to generally treat them the same way. Second, we want to support partial or incorrect code, so we shouldn't care whether or not there is a valid #library directive.

If we really want to limit bolding to only files that contain a #library directive, then we should add API to support this (or change the semantics of the current API). But personally I think the term "library" was being used loosely in that discussion and what was really meant was top-level files (which more closely corresponds to our looser definition). In any case, I think it might be confusing if what we highlight with bold doesn't correspond to the files that we use as analysis roots."

For the first cut, I'm all for going with this specification and am eager to get feedback.  If we want to be more exclusive we can refine as needed.
Review URL: https://chromiumcodereview.appspot.com//9837049

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@5791 260f80e4-7a28-3924-810f-c04153c831b5