Archive timing tests performed to optimize file i/o.
I left in the code to time how long it takes to read from the resulting data structure because I didn't want to forget that we had looked at that as a possible differentiator.

I didn't copy over your code to convert the CharBuffer into a String because I don't think we need to do that. I'm thinking that we'll have one subclass of AbstractScanner that scans from a String and another that scans from a CharBuffer directly.
Review URL: https://chromiumcodereview.appspot.com//10012055

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@6323 260f80e4-7a28-3924-810f-c04153c831b5